Included all of Henrik suggestions and some debug lines added
