+++ package.json
@@ -57,7 +57,6 @@
     "@angular-devkit/core": "^10.2.2",
     "@angular/cli": "^10.2.2",
     "@angular/compiler-cli": "^10.2.4",
-    "@angular/localize": "^10.2.4",
     "@ngrx/store-devtools": "^9.1.0",
     "@types/jasmine": "^3.4.6",
     "@types/moment-timezone": "^0.5.12",
