+++ src/app/files/files-dashboard/files-dashboard.component.ts
@@ -1,15 +1,15 @@
 import { Component, OnDestroy, OnInit } from "@angular/core";
-import { Column } from "../../shared/column.type";
 import { SciCatDataSource } from "../../shared/services/scicat.datasource";
 import { ScicatDataService } from "../../shared/services/scicat-data-service";
 import { ExportExcelService } from "../../shared/services/export-excel.service";
 import { Router } from "@angular/router";
 import * as moment from "moment";
+import { Column } from "shared/modules/shared-table/shared-table.module";
 
 @Component({
   selector: "app-files-dashboard",
   templateUrl: "./files-dashboard.component.html",
-  styleUrls: ["./files-dashboard.component.css"],
+  styleUrls: ["./files-dashboard.component.scss"],
 })
 export class FilesDashboardComponent implements OnInit, OnDestroy {
   end = moment();
