+++ src/app/help/help/help.component.html
@@ -1,7 +1,7 @@
 <p>
   <mat-card class="help-card">
     <mat-card-header>
-      <mat-card-title>Where is the official documentation ?</mat-card-title>
+      <mat-card-title>Where is the official documentation?</mat-card-title>
     </mat-card-header>
     This is our <a href="https://scicatproject.github.io/" target="blank">documentation homepage</a>. 
     Following the documentation link (blue button at top left) you will find detailed information for
