+++ src/app/jobs/jobs-dashboard-new/jobs-dashboard-new.component.ts
@@ -1,10 +1,10 @@
 import { Component, OnDestroy, OnInit } from "@angular/core";
-import { Column } from "../../shared/column.type";
 import { SciCatDataSource } from "../../shared/services/scicat.datasource";
 import { ScicatDataService } from "../../shared/services/scicat-data-service";
 import { ExportExcelService } from "../../shared/services/export-excel.service";
 import { Job } from "shared/sdk";
 import { Router } from "@angular/router";
+import { Column } from "shared/modules/shared-table/shared-table.module";
 
 @Component({
   selector: "app-jobs-new-dashboard",
