+++ src/app/shared/modules/shared-table/shared-table.module.ts
@@ -19,7 +19,7 @@ import { MatCardModule } from "@angular/material/card";
 import { MatCheckboxModule } from "@angular/material/checkbox";
 import { MatIconModule } from "@angular/material/icon";
 import { MatPaginatorModule } from "@angular/material/paginator";
-import { MatSortModule } from "@angular/material/sort";
+import { MatSortModule, SortDirection  } from "@angular/material/sort";
 import { MatTableModule } from "@angular/material/table";
 import { MatToolbarModule } from "@angular/material/toolbar";
 import { MatFormFieldModule } from "@angular/material/form-field";
@@ -32,6 +32,21 @@ import { RouterModule } from "@angular/router";
 import { SatDatepickerModule, SatNativeDateModule } from "saturn-datepicker";
 import { MatDatepickerModule } from "@angular/material/datepicker";
 import { MatProgressSpinnerModule } from "@angular/material/progress-spinner";
+
+export interface Column {
+    id: string;
+    type?: string;
+    visible?: boolean;
+    label: string;
+    hideOrder: number;
+    width?: number;
+    canSort?: boolean;
+    matchMode?: string;
+    format?: string;
+    icon?: string;
+    sortDefault?: SortDirection;
+    filterDefault?: any;
+}
 
 @NgModule({
   declarations: [SharedTableComponent],
